market-arbitrage-scanner/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
urllib3>=1.26.0

# Translation (free)
deep-translator==1.11.4

# Google Trends (free search volume)
pytrends==4.9.2

# CLI interface
click==8.1.7

# Output formatting
tabulate==0.9.0

# Data handling
python-dotenv==1.0.1

# Rate limiting
ratelimit==2.2.1

# User agent rotation
fake-useragent==1.4.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

